Separate extension function from GesturesPluginImpl.
